Remove search module
Remove the Search module im lib and moved the function into the main module. It is now less confusing.
